// osa9/courses-react/src/types.ts
interface CoursePartBase {
    name: string;
    exerciseCount: number;
  }

interface CoursePartBaseWithDescription extends CoursePartBase {
    description: string;
  }
  
  interface CoursePartOne extends CoursePartBaseWithDescription {
    name: "Fundamentals";
  }
  
  interface CoursePartTwo extends CoursePartBase {
    name: "Using props to pass data";
    groupProjectCount: number;
  }
  
  interface CoursePartThree extends CoursePartBaseWithDescription {
    name: "Deeper type usage";
    exerciseSubmissionLink: string;
  }

  interface CoursePartFour extends CoursePartBaseWithDescription {
    name: "Extracurricular type exercise";
  }
  
  export type CoursePart = CoursePartOne | CoursePartTwo | CoursePartThree | CoursePartFour;